refactor(admin): move management view to GestionContainer (FSD) and fix build typescript arity errors
